[MOV] whatsapp: rename to whatsapp_evolution for clarity
Distinguishes the Evolution API-based plugin from the new
whatsapp_api (Meta Cloud API) plugin. DB schema stays "whatsapp"
for backward compatibility. platform stays "whatsapp".
