+++ plugins/whatsapp_evolution/portal/templates/me/whatsapp_authorized.html
